Factor out TopicSettings and ProducerTopic
